(search) Start moving API abstraction to its own layer directory
