+++ applications/search/src/pages/datasets-report-page/report-stats/report-stats.component.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { ReportStats } from './report-stats.component';
 import aggregateDataset from '../../../../test/fixtures/aggregateDataset';
-import { extractStats } from './resolved-report-stats';
+import { extractStats } from '../../../api/get-dataset-stats';
 
 let defaultProps;
 let wrapper;
